+++ src/shared/util/optional.ts
